+++ src/frontend/windows/splash_screen.py
@@ -68,6 +68,7 @@ class SplashScreenLoader(QThread):
         plugin_loader = PluginLoader(self.update_splash_msg)
         plugins = plugin_loader.load_plugins()
         self.config.loaded_plugins.update(plugins)
+        self._update_jinja2_engine_with_plugins()
 
         # check if we have missing keys and remove them from the running config
         plugins = self.config.loaded_plugins.keys()
@@ -96,6 +97,18 @@ class SplashScreenLoader(QThread):
             ),
         }
         self.config.loaded_plugins.update(built_in_plugins)
+
+    def _update_jinja2_engine_with_plugins(self) -> None:
+        for plugin in self.config.loaded_plugins.values():
+            jinja2_filters = getattr(plugin, "jinja2_filters", None)
+            jinja2_functions = getattr(plugin, "jinja2_functions", None)
+            if jinja2_filters:
+                for name, func in jinja2_filters.items():
+                    self.config.jinja_engine.add_filter(name, func)
+
+            if jinja2_functions:
+                for name, func in jinja2_functions.items():
+                    self.config.jinja_engine.add_global(name, func)
 
 
 class SplashScreen(QWidget):